aspy_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Add specific version
stripe==7.7.0
razorpay==1.4.2
python-multipart==0.0.6
redis==4.5.4
reportlab==4.0.7
httpx==0.25.2
